Update any stand-alone ascii character to be parsed correctly.
If a an ascii char was given within a some longer sentence, then we don't handle it. We update ascii char iff they are given as stand-alone chars. I.e. if a char was given as '\x2A', python would read this as '\\x2A' string instead of it actual meaning (it means asterisk). For such cases we now handle these appropriately and translate them to their correct meaning.